requirements.txt
Flask
numpy
torch
transformers
Pillow
